+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 1.0.0 (IN PROGRESS)
 
 * [UIMS-2](https://folio-org.atlassian.net/browse/UIMS-2) Create `ui-mosaic-settings` repository.
+* [UIMS-1](https://folio-org.atlassian.net/browse/UIMS-1) Add "Configuration options" page to settings.